Add benchmarks and changed np.float128 to np.float64

+++ waggon/surrogates/gp.py
@@ -20,19 +20,19 @@ class GP(Surrogate):
             if self.kernel is None:
                 self.kernel = GPy.kern.Matern32(input_dim=X.shape[-1], lengthscale=1.0)
 
-            self.model = GPy.models.GPRegression(X.astype(np.float128), y.astype(np.float128),
+            self.model = GPy.models.GPRegression(X.astype(np.float64), y.astype(np.float64),
                                                  kernel = self.kernel,
                                                  mean_function = self.mean)
         self.mu, self.std = np.mean(y), np.std(y)
         y = (y - self.mu) / (self.std + 1e-8)
         
-        self.model.set_XY(X=X.astype(np.float128) , Y=y.astype(np.float128))
+        self.model.set_XY(X=X.astype(np.float64) , Y=y.astype(np.float64))
         
         self.model.optimize(optimizer='lbfgsb')
 
     def predict(self, X):
 
-        f, var = self.model.predict(X.astype(np.float128))
+        f, var = self.model.predict(X.astype(np.float64))
         std = np.sqrt(var)
 
         f += self.mu
